+++ src/main/java/NumbersProcessor/Main.java
@@ -5,9 +5,8 @@ import NumbersProcessor.logic.NumbersProcessor;
 
 public class Main {
 
-
     public static void main(String[] args) {
-        String path = "d:\\luk\\java\\workspace\\InteliiJ\\1000PrimeNumbers\\src\\main\\resources\\liczbyM.txt";
+        String path = "d:\\luk\\java\\workspace\\InteliiJ\\1000PrimeNumbers\\src\\main\\resources\\liczby.txt";
         Program program = new Program(new NumbersProcessor(), new FileHelper());
         program.startApp(path);
     }
