+++ src/test/java/NumbersProcessor/logic/NumbersProcessorTest.java
@@ -44,6 +44,7 @@ public class NumbersProcessorTest {
         }
         return resultList;
     }
+
     @org.junit.Test
     public void filterNumberStringsTest() {
